Autenticação com spring security

+++ src/main/java/br/com/alura/forum/controller/TopicoController.java
@@ -1,13 +1,17 @@
 package br.com.alura.forum.controller;
 
 import java.net.URI;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -38,21 +43,24 @@ public class TopicoController {
 	private CursoRepository cursoRepository;
 
 	@GetMapping
-	public List<TopicoDTO> listar(String cursoNome) {
-
-		List<Topico> topicos = new ArrayList<>();
+	@Cacheable(value = "listaDeTopicos")
+	public Page<TopicoDTO> listar(@RequestParam(required = false) String cursoNome,
+			@PageableDefault(sort = "id") Pageable paginacao) {
 
 		if (cursoNome == null) {
-			topicos = topicoRepository.findAll();
+			Page<Topico> topicos = topicoRepository.findAll(paginacao);
+			return TopicoDTO.converter(topicos);
 
 		} else {
-			topicos = topicoRepository.findByCursoNome(cursoNome);
+			Page<Topico> topicos = topicoRepository.findByCursoNome(cursoNome, paginacao);
+			return TopicoDTO.converter(topicos);
 		}
 
-		return TopicoDTO.converter(topicos);
 	}
 
 	@PostMapping
+	@Transactional
+	@CacheEvict(value = "listaDeTopicos", allEntries = true)
 	public ResponseEntity<TopicoDTO> cadastrar(@RequestBody @Valid TopicosForm form, UriComponentsBuilder uriBuilder) {
 
 		Topico topico = form.converter(cursoRepository);
@@ -66,28 +74,48 @@ public class TopicoController {
 	}
 
 	@GetMapping("{id}")
-	public DetalhesdoTopicoDTO detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
+	public ResponseEntity<DetalhesdoTopicoDTO> detalhar(@PathVariable Long id) {
 
-		return new DetalhesdoTopicoDTO(topico);
+		Optional<Topico> topico = topicoRepository.findById(id);
+
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesdoTopicoDTO(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
 
 	@PutMapping("{id}")
 	@Transactional
+	@CacheEvict(value = "listaDeTopicos", allEntries = true)
 	public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicosForm form) {
 
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optionalTopico = topicoRepository.findById(id);
 
-		return ResponseEntity.ok(new TopicoDTO(topico));
+		if (optionalTopico.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDTO(topico));
+		}
+
+		return ResponseEntity.notFound().build();
 
 	}
 
 	@DeleteMapping("/{id}")
+	@Transactional
+	@CacheEvict(value = "listaDeTopicos", allEntries = true)
 	public ResponseEntity<?> remover(@PathVariable Long id) {
 
-		topicoRepository.deleteById(id);
+		Optional<Topico> optionalTopico = topicoRepository.findById(id);
 
-		return ResponseEntity.ok().build();
+		if (optionalTopico.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
 
 }
